+++ NEWS.md
@@ -1,4 +1,5 @@
 # glamr 1.0.1
+* Resolve bug with `resolve_known_issues` that resulted from having both targets and results for the same mechanism and indicator that need to be removed.
 * Add set of functions to store, access, and load Panorama credentials - `set_pano`,
  `pano_user()`, `pano_pwd()`, `pano_session()`, and `load_secrets()`.
 * Add functions to discover and download output datasets from Panorama: `pano_content()`, `pano_elements()`, `pano_unpack()`, `pano_extract()`, `pano_download()`, `pano_items()`, `pano_extract_msd()`